feat(android): add File object wrapper for supporting URI based paths
Most popular RN 3rd party libraries handling files (e.g. image/photo handling)
are returning file:// URIs on Android. To save handling this in JS several
times, we can simply try to detect if we are handling an URI and parse them
accordingly. Regardless of the input we always return an File object to handle
files.

Signed-off-by: Martin Mazein <internet+github@mazein.net>